moved enhanced_mean_image()
