+++ suite2p/extraction/__init__.py
@@ -1,3 +1,3 @@
 from .dcnv import preprocess, oasis
-from .extract import extract_traces, detect_and_extract
+from .extract import extract_traces, detect_and_extract, enhanced_mean_image
 from .masks import create_cell_masks, create_neuropil_masks, roi_stats, get_overlaps
